+++ lib/ESP_WiFiManager_Lite/examples/ESP_WiFi/Credentials.h
@@ -2,8 +2,8 @@
   Credentials.h
   For ESP8266 / ESP32 boards
 
-  ESP_WiFiManager_Lite (https://github.com/khoih-prog/ESP_WiFiManager_Lite) is a library 
-  for the ESP32/ESP8266 boards to enable store Credentials in EEPROM/SPIFFS/LittleFS for easy 
+  ESP_WiFiManager_Lite (https://github.com/khoih-prog/ESP_WiFiManager_Lite) is a library
+  for the ESP32/ESP8266 boards to enable store Credentials in EEPROM/SPIFFS/LittleFS for easy
   configuration/reconfiguration and autoconnect/autoreconnect of WiFi and other services without Hardcoding.
 
   Built by Khoi Hoang https://github.com/khoih-prog/ESP_WiFiManager_Lite
@@ -15,11 +15,11 @@
 
 #include "defines.h"
 
-/// Start Default Config Data //////////////////
+/// Start Default Config Data ///
 
 /*
 #define SSID_MAX_LEN      32
-//From v1.0.3, WPA2 passwords can be up to 63 characters long.
+// WPA2 passwords can be up to 63 characters long.
 #define PASS_MAX_LEN      64
 
 typedef struct
@@ -30,9 +30,9 @@ typedef struct
 
 #define NUM_WIFI_CREDENTIALS      2
 
-// Configurable items besides fixed Header, just add board_name 
+// Configurable items besides fixed Header, just add board_name
 #define NUM_CONFIGURABLE_ITEMS    ( ( 2 * NUM_WIFI_CREDENTIALS ) + 1 )
-////////////////
+
 
 typedef struct Configuration
 {
@@ -48,7 +48,7 @@ typedef struct Configuration
 #if TO_LOAD_DEFAULT_CONFIG_DATA
 
 // This feature is primarily used in development to force a known set of values as Config Data
-// It will NOT force the Config Portal to activate. Use DRD or erase Config Data with ESP_WiFiManager.clearConfigData()
+// It will NOT force the Config Portal to activate. Use MRD or erase Config Data with ESP_WiFiManager.clearConfigData()
 
 // Used mostly for development and debugging. FORCES default values to be loaded each run.
 // Config Portal data input will be ignored and overridden by DEFAULT_CONFIG_DATA
@@ -62,7 +62,7 @@ bool LOAD_DEFAULT_CONFIG_DATA = false;
 ESP_WM_LITE_Configuration defaultConfig =
 {
   //char header[16], dummy, not used
-#if ESP8266 
+#if ESP8266
   "ESP8266",
 #else
   "ESP32",
@@ -74,7 +74,7 @@ ESP_WM_LITE_Configuration defaultConfig =
   "SSID2",  "password2",
   //char board_name     [24];
 
-#if ESP8266 
+#if ESP8266
   "ESP8266-Control",
 #else
   "ESP32-Control",